+++ src/SharedVariables.ts
@@ -26,6 +26,7 @@ export class SharedVariables implements vscode.Disposable {
 		public execWatchTimeout: number,
 		private _execRunningTimeout: null | number,
 		public isNoThrow: boolean,
+		public defaultEnv: { [prop: string]: string }
 	) { }
 
 	dispose() {