+++ src/test/FsWrapper.test.ts
@@ -11,7 +11,7 @@ import * as vscode from 'vscode';
 import { spawnAsync, statAsync } from '../FsWrapper';
 import { ChildProcessStub } from './TestCommon';
 
-describe('FsWrapper.spawnAsync', function () {
+describe(path.basename(__filename), function () {
   it('echoes', async function () {
     const isWin = process.platform === 'win32';
     const opt: SpawnOptions = isWin ? { shell: true } : {};
@@ -20,17 +20,6 @@ describe('FsWrapper.spawnAsync', function () {
     assert.equal(r.output.length, 2);
     assert.equal(r.output[0], 'apple' + EOL);
     assert.equal(r.status, 0);
-  })
-
-  it.skip('sleeps', async function () {
-    this.timeout(1100);
-    this.slow(1050);
-    if (process.platform === 'darwin') {
-      const r = await spawnAsync('sleep', ['1']);
-      assert.equal(r.stdout, '');
-      assert.equal(r.output.length, 0);
-      assert.equal(r.status, 0);
-    }
   })
 
   it('not existing', function () {
@@ -152,13 +141,16 @@ describe('ChildProcessStub', function () {
     });
   })
 
-  it.skip('should works2', async function () {
+  it('should works2', async function () {
+    this.timeout(700);
+    this.slow(600);
     const cp = new ChildProcessStub();
     let output: string = '';
     cp.stdout.on('data', (d: string) => {
       output += d;
     });
     cp.write('alma');
+    setTimeout(() => { cp.close(); }, 500);
     await new Promise(resolve => {
       cp.on('close', resolve);
     }).then(() => {